requirements.txt
prometheus_client
kafka
clickhouse-driver
python-dotenv
pytz